+++ .credo.exs
@@ -73,7 +73,7 @@
           ## Design Checks
           #
           # You can customize the priority of any check
-          # Priority values are: `low, normal, high, higher`
+          # Priority examples are: `low, normal, high, higher`
           #
           {Credo.Check.Design.AliasUsage,
            [priority: :low, if_nested_deeper_than: 2, if_called_more_often_than: 0]},