+++ test/phx_live_storybook/entries_validator_test.exs
@@ -402,6 +402,35 @@ defmodule PhxLiveStorybook.EntriesValidatorTest do
     end
   end
 
+  describe "attribute examples is a list and is matching declared type" do
+    test "with an empty list it wont raise" do
+      entry = entry_with_attr(id: :attr, type: :atom, examples: [])
+      assert validate(entry)
+    end
+
+    test "with a list of matching type, it wont raise" do
+      entry = entry_with_attr(id: :attr, type: :atom, examples: [:foo, :bar])
+      assert validate(entry)
+    end
+
+    test "with a list of integer and a range, it wont raise" do
+      entry = entry_with_attr(id: :attr, type: :integer, examples: 1..10)
+      assert validate(entry)
+    end
+
+    test "without a list, it will raise" do
+      entry = entry_with_attr(id: :attr, type: :atom, examples: :foo)
+      e = assert_raise CompileError, fn -> validate(entry) end
+      assert e.description =~ "examples for attr :attr must be a list of :atom"
+    end
+
+    test "with a list of non matching type, it will raise" do
+      entry = entry_with_attr(id: :attr, type: :atom, examples: ["foo", "bar"])
+      e = assert_raise CompileError, fn -> validate(entry) end
+      assert e.description =~ "examples for attr :attr must be a list of :atom"
+    end
+  end
+
   describe "attribute values is a list and is matching declared type" do
     test "with an empty list it wont raise" do
       entry = entry_with_attr(id: :attr, type: :atom, values: [])
@@ -431,40 +460,13 @@ defmodule PhxLiveStorybook.EntriesValidatorTest do
     end
   end
 
-  describe "attribute values! is a list and is matching declared type" do
-    test "with an empty list it wont raise" do
-      entry = entry_with_attr(id: :attr, type: :atom, values!: [])
-      assert validate(entry)
-    end
-
-    test "with a list of matching type, it wont raise" do
-      entry = entry_with_attr(id: :attr, type: :atom, values!: [:foo, :bar])
-      assert validate(entry)
-    end
-
-    test "with a list of integer and a range, it wont raise" do
-      entry = entry_with_attr(id: :attr, type: :integer, values!: 1..10)
-      assert validate(entry)
-    end
-
-    test "without a list, it will raise" do
-      entry = entry_with_attr(id: :attr, type: :atom, values!: :foo)
-      e = assert_raise CompileError, fn -> validate(entry) end
-      assert e.description =~ "values! for attr :attr must be a list of :atom"
-    end
-
-    test "with a list of non matching type, it will raise" do
-      entry = entry_with_attr(id: :attr, type: :atom, values!: ["foo", "bar"])
-      e = assert_raise CompileError, fn -> validate(entry) end
-      assert e.description =~ "values! for attr :attr must be a list of :atom"
-    end
-  end
-
-  describe "attribute values & values! cant be set at the same time" do
+  describe "attribute examples & values cant be set at the same time" do
     test "it raises" do
-      entry = entry_with_attr(id: :attr, type: :atom, values: [:foo, :var], values!: [:foo, :bar])
+      entry =
+        entry_with_attr(id: :attr, type: :atom, examples: [:foo, :var], values: [:foo, :bar])
+
       e = assert_raise CompileError, fn -> validate(entry) end
-      assert e.description =~ "values and values! for attr :attr cannot be set at the same time"
+      assert e.description =~ "examples and values for attr :attr cannot be set at the same time"
     end
   end
 
@@ -725,18 +727,18 @@ defmodule PhxLiveStorybook.EntriesValidatorTest do
       assert e.description =~ "attribute :bar in story :foo, group :group must be of type: :atom"
     end
 
-    test "story attribute value matching attribute values! wont raise" do
+    test "story attribute value matching attribute values wont raise" do
       entry = %ComponentEntry{
-        attributes: [%Attr{id: :attr, type: :atom, values!: [:foo, :bar]}],
+        attributes: [%Attr{id: :attr, type: :atom, values: [:foo, :bar]}],
         stories: [%Story{id: :foo, attributes: %{attr: :bar}}]
       }
 
       assert validate(entry)
     end
 
-    test "story attribute value don't matching attribute values! will raise" do
+    test "story attribute value don't matching attribute values will raise" do
       entry = %ComponentEntry{
-        attributes: [%Attr{id: :attr, type: :atom, values!: [:foo, :bar]}],
+        attributes: [%Attr{id: :attr, type: :atom, values: [:foo, :bar]}],
         stories: [%Story{id: :foo, attributes: %{attr: :not_matching}}]
       }
 
@@ -744,9 +746,9 @@ defmodule PhxLiveStorybook.EntriesValidatorTest do
       assert e.description =~ "attribute :attr in story :foo must be one of [:foo, :bar]"
     end
 
-    test "story group attribute value don't matching attribute values! will raise" do
+    test "story group attribute value don't matching attribute values will raise" do
       entry = %ComponentEntry{
-        attributes: [%Attr{id: :attr, type: :atom, values!: [:foo, :bar]}],
+        attributes: [%Attr{id: :attr, type: :atom, values: [:foo, :bar]}],
         stories: [
           %StoryGroup{id: :group, stories: [%Story{id: :foo, attributes: %{attr: :not_matching}}]}
         ]
@@ -758,9 +760,9 @@ defmodule PhxLiveStorybook.EntriesValidatorTest do
                "attribute :attr in story :foo, group :group must be one of [:foo, :bar]"
     end
 
-    test "story attribute value don't matching attribute values won't raise" do
+    test "story attribute value don't matching attribute examples won't raise" do
       entry = %ComponentEntry{
-        attributes: [%Attr{id: :attr, type: :atom, values: [:foo, :bar]}],
+        attributes: [%Attr{id: :attr, type: :atom, examples: [:foo, :bar]}],
         stories: [%Story{id: :foo, attributes: %{attr: :not_matching}}]
       }
 
@@ -913,8 +915,8 @@ defmodule PhxLiveStorybook.EntriesValidatorTest do
           doc: opts[:doc],
           default: opts[:default],
           required: opts[:required],
-          values: opts[:values],
-          values!: opts[:values!]
+          examples: opts[:examples],
+          values: opts[:values]
         }
       ]
     }