+++ test/phx_live_storybook/live/playground_live_test.exs
@@ -361,7 +361,7 @@ defmodule PhxLiveStorybook.PlaygroundLiveTest do
       form_label_selector = "#tree_storybook_template_component-playground-form_label"
       form_toggle_selector = "#tree_storybook_template_component-playground-form_status"
 
-      assert get_element_attribute(view, form_label_selector, "value") == "[Multiple values]"
+      assert get_element_attribute(view, form_label_selector, "value") == "[Multiple examples]"
 
       playground_preview_view |> element("#two #set-foo") |> render_click()
       assert render(playground_element) =~ "template_component: foo / status: false"
@@ -374,14 +374,14 @@ defmodule PhxLiveStorybook.PlaygroundLiveTest do
       assert render(playground_element) =~ "template_component: bar / status: true"
       assert render(playground_element) =~ "template_component: bar / status: false"
 
-      assert get_element_attribute(view, form_toggle_selector, "value") == "[Multiple values]"
+      assert get_element_attribute(view, form_toggle_selector, "value") == "[Multiple examples]"
       playground_preview_view |> element("#two #toggle-status") |> render_click()
       view |> form(form_selector, %{playground: %{status: "true"}}) |> render_change()
       assert render(playground_element) =~ "template_component: bar / status: true"
       refute render(playground_element) =~ "template_component: bar / status: false"
     end
 
-    test "playground with template values", %{conn: conn} do
+    test "playground with template examples", %{conn: conn} do
       {:ok, view, _html} =
         live(
           conn,